# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DB API — Multitenant REST API",
        "description": "Laravel 13 API with DDD/CQRS architecture and table-per-tenant multi-tenancy.\n\n## Authentication\nAll write endpoints require a Bearer token issued by Laravel Passport.\nObtain one via `POST /oauth/token` (standard Passport password grant).\n\n## Multitenancy\nEvery request must target a tenant via the URL `/tenant_id/v1/...`.\nTenant identifiers are 3-63 lowercase alphanumeric characters (hyphens/underscores allowed).\n\n## Module gating\nSome endpoints require the tenant to have a specific module enabled.\nIf a module is disabled, the API returns **403** with the message\n`\"Module 'X' is not enabled for this tenant\"`.\n\n## Standard response envelope\nAll endpoints return a consistent JSON envelope:\n```json\n{\n  \"success\": true,\n  \"data\": { ... },\n  \"message\": \"Operation completed successfully\"\n}\n```\nErrors replace `data` with `errors` (validation details).",
        "contact": {
            "email": "dev@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local development"
        },
        {
            "url": "https://api.example.com",
            "description": "Production"
        }
    ],
    "paths": {
...
        "/{tenant}/{version}/pages/{id}": {
            "get": {
                "tags": [
                    "Pages"
                ],
                "summary": "Get a page by ID in a specific language",
                "description": "Retrieve a page in the requested language. If language translation doesn't exist, returns first available translation.",
                "operationId": "d16f1411ef82537a10daa8549dba0cce",
                "parameters": [
                    {
                        "name": "tenant",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "acme"
                    },
                    {
                        "name": "version",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "v1"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "lang",
                        "in": "query",
                        "description": "ISO 639-1 language code (defaults to 'en')",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "en"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Page found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Pages"
                ],
                "summary": "Update a page or add/update a translation",
                "description": "Update page content for a specific language. If translation exists, updates it; if not, creates new translation. Can also update page status (core field).",
                "operationId": "419f036d36898d2c630c739f84e22295",
                "parameters": [
                    {
                        "name": "tenant",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "acme"
                    },
                    {
                        "name": "version",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "v1"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "language_code"
                                ],
                                "properties": {
                                    "language_code": {
                                        "type": "string",
                                        "example": "es"
                                    },
                                    "status": {
                                        "type": "string",
                                        "nullable": true,
                                        "enum": [
                                            "draft",
                                            "published",
                                            "archived"
                                        ]
                                    },
                                    "slug": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "title": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "content": {
                                        "type": "array",
                                        "items": {
                                            "type": "object"
                                        },
                                        "nullable": true
                                    },
                                    "seo_title": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "seo_description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "seo_keywords": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "canonical_url": {
                                        "type": "string",
                                        "format": "url",
                                        "nullable": true
                                    },
                                    "og_title": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "og_description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "og_image": {
                                        "type": "string",
                                        "format": "url",
                                        "nullable": true
                                    },
                                    "structured_data": {
                                        "type": "object",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Page updated"
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "description": "Standard error response envelope",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Validation failed"
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "title": [
                                "The title field is required."
                            ]
                        },
                        "nullable": true,
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
            "PageResponse": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "draft",
                            "published"
                        ]
                    },
                    "language_code": {
                        "type": "string",
                        "example": "en"
                    },
                    "title": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "content": {
                        "description": "Block editor JSON or HTML",
                        "type": "string",
                        "nullable": true
                    },
                    "seo_title": {
                        "type": "string",
                        "nullable": true
                    },
                    "seo_description": {
                        "type": "string",
                        "nullable": true
                    },
                    "canonical_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "og_title": {
                        "type": "string",
                        "nullable": true
                    },
                    "og_description": {
                        "type": "string",
                        "nullable": true
                    },
                    "og_image": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Passport",
                "description": "Laravel Passport token. Obtain via POST /oauth/token (password grant)."
            }
        }
    },
...
}